add handleFormSubmit.js in utils and render.js in components/App
