+++ js-structure/js-structure_quiz-app/components/App/render.js
@@ -1,0 +1,6 @@
+import App from "./App";
+
+export default function render() {
+    const root = document.querySelector("#root");
+    root.append(App());
+  }
